chinook
In this project I've used chinook database that contains information about the artists, songs, and albums from the music shop, as well as information on the shop's employees, customers, and the customers purchases. This information is contained in eleven tables. Using this data I've analyzed Employee Sales Performance, sales by country, Top albums based on Genres.